+++ wiki/Changelog.md
@@ -1,3 +1,4 @@
+  * Feature to search scrollback buffer (#85); shortcuts Alt+F3 or Shift+Ctrl+H; configuration options.
   * New option ZoomFontWithWindow to disable Shift-coupled font-with-window zooming (#476).
   * New OSC sequence to copy the window title to the clipboard (#303).